backend/pyproject.toml
[build-system]
requires = ["poetry==1.1.4"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "knowledge-space"
description = ""
version = "0.0.1"
readme = "README.md"
authors = ['Nikita Kurushin', 'Roman Krasnov', 'Sergay Deev']

[tool.poetry.dependencies]
python = "3.9.2"

django = "3.1.*"
django-cors-headers = "^3.7.0"
django-split-settings = "^1.0.1"
python-decouple = "^3.4"
djangorestframework-simplejwt = "^4.6.0"
djangorestframework = "^3.12.2"
psycopg2-binary = "^2.8.6"
drf-yasg = "^1.20.0"
djangorestframework-camel-case = "^1.2.0"
django-filter = "^2.4.0"
pre-commit = "^2.12.0"
black = "^20.8b1"
isort = "^5.8.0"
autoflake = "^1.4"
flake8 = "^3.9.0"
drf-spectacular = "^0.16.0"
django-phone-field = "^1.8.1"
Faker = "^8.4.0"
gunicorn = "^20.1.0"
